restaurant-service/requirements.txt
flask==2.3.3
pymongo==4.5.0
PyJWT==2.8.0
Werkzeug==2.3.7
